requirements.txt
openai-whisper==20231117
python-dotenv==1.0.0
torch>=2.0.0
